認証チェック用のProxyを追加
